+++ modules/bridge/src/main/ts/ephox/bridge/core/MenuButton.ts
@@ -38,7 +38,6 @@ export interface BaseMenuButtonInstanceApi {
   setActive: (state: boolean) => void;
   setText: (text: string) => void;
   setIcon: (icon: string) => void;
-  setTooltip: (tooltip: string) => void;
 }
 
 export const baseMenuButtonFields = [